+++ components/board/BestArticle.tsx
@@ -3,16 +3,13 @@ import styles from "@/styles/BestArticle.module.scss";
 import Image from "next/image";
 import Link from "next/link";
 import { getDate } from "@/utils/getDate";
-import useDeviceSize from "@/hooks/useDeviceSize";
-import { useEffect, useState } from "react";
-import axios from "@/lib/axios";
 
 // NOTE - 베스트 게시글
 interface BestArticleProps {
   article: Article;
 }
 
-function BestArticleCard({ article }: BestArticleProps) {
+export default function BestArticleCard({ article }: BestArticleProps) {
   return (
     <Link href={`addboard/${article.id}`}>
       <article className={styles["best-article-container"]}>
@@ -55,43 +52,5 @@ function BestArticleCard({ article }: BestArticleProps) {
         </div>
       </article>
     </Link>
-  );
-}
-
-// NOTE - 베스트 게시글 리스트
-
-export default function BestArticleList() {
-  const { isDesktop, isMobile, isTablet, bestPageSizeCount } = useDeviceSize();
-  const [bestArticleList, setBestArticleList] = useState<Article[]>([]);
-  const [bestPageSize, setBestPageSize] = useState(bestPageSizeCount);
-
-  async function getBestArticleList() {
-    const res = await axios.get(
-      `/articles?pageSize=${bestPageSize}&orderBy=like`
-    );
-    const nextArticleList = res.data.list;
-    setBestArticleList(nextArticleList);
-  }
-
-  useEffect(() => {
-    getBestArticleList();
-  }, [bestPageSize]);
-
-  useEffect(() => {
-    if (isMobile) {
-      setBestPageSize(1);
-    } else if (isTablet) {
-      setBestPageSize(2);
-    } else if (isDesktop) {
-      setBestPageSize(3);
-    }
-  }, [isMobile, isTablet, isDesktop]);
-
-  return (
-    <div className={styles["best-article-list-container"]}>
-      {bestArticleList.map((article) => {
-        return <BestArticleCard key={article.id} article={article} />;
-      })}
-    </div>
   );
 }
